- consolidated erd enums into single file and boolified
