+++ gehomesdk/erd/converters/laundry/machine_state_converter.py
@@ -12,4 +12,4 @@ class MachineStateConverter(ErdReadOnlyConverter[MachineState]):
             om = ErdMachineState(erd_decode_int(value))
             return MACHINE_STATE_MAP[om].value
         except (KeyError, ValueError):
-            return ErdMachineState.NA
+            return ErdMachineState.IDLE
