+++ gehomesdk/erd/values/laundry/__init__.py
@@ -1,49 +1,33 @@
-from .erd_machine_state import ErdMachineState
+from .laundry_enums import *
 from .machine_state import MachineState
 from .machine_state_mapping import MACHINE_STATE_MAP
-from .erd_laundry_door_status import ErdLaundryDoorStatus
 from .laundry_door_status import LaundryDoorStatus
 from .laundry_door_status_mapping import LAUNDRY_DOOR_STATUS_MAP
-from .erd_laundry_door_lock_status import ErdLaundryDoorLockStatus
 from .laundry_door_lock_status import LaundryDoorLockStatus
 from .laundry_door_lock_status_mapping import LAUNDRY_DOOR_LOCK_STATUS_MAP
-from .erd_laundry_sub_cycle import ErdLaundrySubCycle
 from .laundry_sub_cycle import LaundrySubCycle
 from .laundry_sub_cycle_mapping import LAUNDRY_SUB_CYCLE_MAP
-from .erd_laundry_cycle import ErdLaundryCycle
 from .laundry_cycle import LaundryCycle
 from .laundry_cycle_mapping import LAUNDRY_CYCLE_MAP
-from .erd_rinse_option import ErdRinseOption
 from .rinse_option import RinseOption
 from .rinse_option_mapping import RINSE_OPTION_MAP
-from .erd_temperature_option import ErdTemperatureOption, ErdTemperatureNewOption
 from .temperature_option import TemperatureOption
 from .temperature_option_mapping import TEMPERATURE_OPTION_MAP, TEMPERATURENEW_OPTION_MAP
-from .erd_washtemp_level import ErdWashTempLevel
 from .washtemp_level import WashTempLevel
 from .washtemp_level_mapping import WASHTEMP_LEVEL_MAP
-from .erd_dryness_level import ErdDrynessLevel, ErdDrynessNewLevel
 from .dryness_level import DrynessLevel
 from .dryness_level_mapping import DRYNESS_LEVEL_MAP, DRYNESSNEW_LEVEL_MAP
-from .erd_tumble_status import ErdTumbleStatus, ErdTumbleStatus
 from .tumble_status import TumbleStatus
 from .tumble_status_mapping import TUMBLE_STATUS_MAP
-from .erd_spintime_level import ErdSpinTimeLevel
 from .spintime_level import SpinTimeLevel
 from .spintime_level_mapping import SPINTIME_LEVEL_MAP
-from .erd_soil_level import ErdSoilLevel
 from .soil_level import SoilLevel
 from .soil_level_mapping import SOIL_LEVEL_MAP
-from .erd_remote_status import ErdRemoteStatus
 from .remote_status import RemoteStatus
 from .remote_status_mapping import REMOTE_STATUS_MAP
-from .erd_tank_status import ErdTankStatus
 from .tank_status import TankStatus
 from .tank_status_mapping import TANK_STATUS_MAP
-from .erd_tank_selected import ErdTankSelected
 from .tank_selected import TankSelected
 from .tank_selected_mapping import TANK_SELECTED_MAP
-from .erd_sheet_usage_configuration import ErdSheetUsageConfiguration
-from .erd_ecodry_status import ErdEcoDryStatus
 from .ecodry_status import EcoDryStatus
 from .ecodry_status_mapping import ECODRY_STATUS_MAP
