--- gehomesdk/erd/values/laundry/erd_ecodry_status.py
@@ -1,8 +1,0 @@
-import enum
-
-@enum.unique
-class ErdEcoDryStatus(enum.Enum):
-    STATE_0 = 0
-    STATE_1 = 1
-    STATE_2 = 2
-    STATE_3 = 3
