--- gehomesdk/erd/values/laundry/erd_laundry_cycle.py
@@ -1,59 +1,0 @@
-import enum
-
-@enum.unique
-class ErdLaundryCycle(enum.Enum):
-    INVALID = 0
-    BASKET_CLEAN = 1
-    DRAIN_AND_SPIN = 2
-    QUICK_RINSE = 3
-    BULKY_ITEMS = 4
-    SANITIZE = 5
-    TOWELS = 6
-    STEAM_REFRESH = 7
-    NORMAL = 8
-    WHITES = 9
-    DARK_COLORS = 10
-    JEANS = 11
-    HAND_WASH = 12
-    DELICATES = 13
-    SPEED_WASH = 14
-    HEAVY_DUTY = 15
-    ALLERGEN = 16
-    POWER_CLEAN = 17
-    RINSE_AND_SPIN = 18
-    SINGLE_ITEM_WASH = 19
-    COLORS = 20
-    COLD_WASH = 21
-    WATER_ON_DEMAND = 22
-    TUB_CLEAN = 23
-    CASUALS_WITH_STEAM = 24
-    STAIN_WASH_WITH_STEAM = 25
-    BULKY_BEDDING = 26
-    NORMAL2 = 28
-    QUICK_WASH = 29
-    SANITIZE_WITH_OXI = 30
-    SELF_CLEAN = 31
-    TOWEL = 32
-    ULTRAFRESH_VENT = 35
-    COTTONS = 128
-    EASY_CARE = 129
-    ACTIVE_WEAR = 130
-    TIMED_DRY = 131
-    DEWRINKLE = 132
-    AIR_FLUFF = 133
-    STEAM_REFRESH2 = 134
-    STEAM_DEWRINKLE = 135
-    SPEED_DRY = 136
-    MIXED = 137
-    QUICK_DRY = 138
-    CASUALS = 139
-    WARM_UP = 140
-    ENERGY_SAVER = 141
-    ANTIBACTERIAL = 142
-    RACK_DRY = 143
-    BABY_CARE = 144
-    AUTO_DRY = 145
-    AUTO_EXTRA_DRY = 146
-    PERM_PRESS = 147
-    WASHER_LINK = 148
-    AUTO_DAMP_DRY = 149
