--- gehomesdk/erd/values/laundry/erd_laundry_door_lock_status.py
@@ -1,6 +1,0 @@
-import enum
-
-@enum.unique
-class ErdLaundryDoorLockStatus(enum.Enum):
-    LOCKED = 1
-    UNLOCKED = 0
