--- gehomesdk/erd/values/laundry/erd_laundry_door_status.py
@@ -1,6 +1,0 @@
-import enum
-
-@enum.unique
-class ErdLaundryDoorStatus(enum.Enum):
-    OPEN = 0
-    CLOSED = 1
