--- gehomesdk/erd/values/laundry/erd_remote_status.py
@@ -1,7 +1,0 @@
-import enum
-
-@enum.unique
-class ErdRemoteStatus(enum.Enum):
-    DISABLE = 0
-    ENABLED = 1
-    NOT_SUPPORTED = 255
