--- gehomesdk/erd/values/laundry/erd_rinse_option.py
@@ -1,10 +1,0 @@
-import enum
-
-@enum.unique
-class ErdRinseOption(enum.Enum):
-    INVALID = 0
-    NORMAL = 1
-    EXTRA = 2
-    MAX = 3
-    HEAVY = 4
-    EXTRA_HEAVY = 5
