--- gehomesdk/erd/values/laundry/erd_tank_selected.py
@@ -1,9 +1,0 @@
-import enum
-
-@enum.unique
-class ErdTankSelected(enum.Enum):
-    DETERGENT = 0
-    SOFTENER = 1
-    BLEACH = 2
-    OTHER = 3
-    INVALID = 4
