--- gehomesdk/erd/values/laundry/erd_tank_status.py
@@ -1,10 +1,0 @@
-import enum
-
-@enum.unique
-class ErdTankStatus(enum.Enum):
-    FULL = 100
-    PERCENT_75 = 75
-    PERCENT_50 = 50
-    PERCENT_25 = 25
-    EMPTY = 0
-
