--- gehomesdk/erd/values/laundry/erd_tumble_status.py
@@ -1,8 +1,0 @@
-import enum
-
-@enum.unique
-class ErdTumbleStatus(enum.Enum):
-    NOT_AVAILABLE = 255
-    ENABLE = 1
-    DISABLE = 0
-
