--- gehomesdk/erd/values/laundry/erd_tumble_type.py
@@ -1,8 +1,0 @@
-import enum
-
-@enum.unique
-class ErdTumbleType(enum.Enum):
-    INVALID = 255
-    LEGACY = 1
-    NEW = 0
-
